+++ CHANGELOG.md
@@ -1,9 +1,16 @@
-## 0.7.2
+## 0.8.+
+* Fixed receipt validation param for `android`.
+* Updated `http` package.
 * Implemented new method `getAppStoreInitiatedProducts`.
   - Handling of iOS method `paymentQueue:shouldAddStorePayment:forProduct:`
   - Has no effect on Android.
 * Fixed issue with method `buyProductWithoutFinishTransaction` for iOS, was not getting the productId.
 * Fixed issue with `toString` method of class `IAPItem`, was printing incorrect values.
+* Fixes for #44. Unsafe getting `originalJson` when restoring item and `Android`.
+* Use dictionaryWithObjectsAndKeys in NSDictionary to fetch product values. This will prevent from NSInvalidArgumentException in ios which rarely occurs.
+* Fixed wrong npe in `android` when `getAvailablePurchases`.
++ Only parse `orderId` when exists in `Android` to prevent crashing.
++ Add additional success purchase listener in `iOS`. Related [#54](https://github.com/dooboolab/flutter_inapp_purchase/issues/54)
 
 ## 0.7.1
 * Implemented receiptValidation for both android and ios.